+++ Changelog.md
@@ -6,9 +6,11 @@ Version 0.13
 ------------
 ### Features ###
 * Music files (m4a and mp3) can be dragged onto the dock icon to go the directory of the file
+* Can drag directory but now allows all files 
 
 ### Bug fixes ###
 * Fixed Warnings about missing method
+* Fixed a bug where no window was open
 
 ------------
 Version 0.12 (Fri Aug 05 2011 02:23:54 +0100)