+++ Changelog.md
@@ -11,7 +11,7 @@ Version 0.13
 * Tags can now do bitrate channels sample rate and kind 
 * FileSystemNode can now tell it size
 * Disc and track pairs
-
+* Prefs for hiding columns
 
 ### Bug fixes ###
 * Fixed Warnings about missing method