+++ Changelog.md
@@ -2,7 +2,13 @@ Changelog
 =========
 
 ------------
-Version 0.13
+Version 0.14
+------------
+### Features ###
+
+
+------------
+Version 0.13 (Sun Aug 07 2011 23:39:44 +0100)
 ------------
 ### Features ###
 * Music files (m4a and mp3) can be dragged onto the dock icon to go the directory of the file
@@ -19,6 +25,7 @@ Version 0.13
 ### Bug fixes ###
 * Fixed Warnings about missing method
 * Fixed a bug where no window was open
+
 
 ------------
 Version 0.12 (Fri Aug 05 2011 02:23:54 +0100)